Move App and Document to src
